refactor: component/tab 폴더 구조 변경 (/tab)
컴파운드 패턴으로 조합 가능한 ui를 만듦에 따라서 도메인을 결합한 추상화 단계를 둘 필요성이 존재하였다.

이에 도메인 결함한 단계를 파일 최상단에, 내부 구현 로직을 하위에 두도록 하였다.

추가적으로 도메인과 관련한 상수데이터 및 타입은 외부에서 공통으로 사용하지 않는다면 한 파일 내에서 관리하도록 하였다.

#74
