+++ FE/src/components/common/tabs/tab/TabCompound/TabCompound.tsx
