+++ FE/src/components/main/Program.compound.tsx
@@ -10,8 +10,8 @@ import {
 } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 import ErrorFallback from "../common/ErrorFallback";
-import Tab from "../common/tabs/Tab";
-import TextTab from "../common/tabs/TextTab";
+import Tab from "../common/tabs/tab/Tab";
+import TextTab from "../common/tabs/tab/TextTab";
 import ProgramList from "./ProgramList";
 import ProgramListLoader from "./ProgramList.loader";
 import MAIN from "@/constants/MAIN";
